Spliting dispatcher from worker: renaming worker -> dispatcher
