+++ dispatcher/joulupukki/dispatcher/lib/packer.py
