+++ dispatcher/joulupukki/dispatcher/lib/rpmpacker.py
